hir_expand::attrs

Function collect_item_tree_attrs

pub fn collect_item_tree_attrs<'a, BreakValue>(
    owner: &dyn HasAttrs,
    cfg_options: impl Fn() -> &'a CfgOptions,
    on_attr: impl FnMut(Meta, &SyntaxNode, &Attr, TextRange) -> ControlFlow<BreakValue>,
) -> Option<Either<BreakValue, CfgExpr>>

This collects attributes exactly as the item tree needs them. This is used for the item tree, as well as for resolving AttrIds.