hir_expand::attrs

Function expand_cfg_attr

pub fn expand_cfg_attr<'a, BreakValue>(
    attrs: impl Iterator<Item = Attr>,
    cfg_options: impl FnMut() -> &'a CfgOptions,
    callback: impl FnMut(Meta, Attr) -> ControlFlow<BreakValue>,
) -> Option<BreakValue>

The callback is passed the attribute and the outermost ast::Attr. Note that one node may map to multiple [ast::Meta]s due to cfg_attr.

unsafe(attr) are passed the inner attribute for now.