hir_expand::builtin

Module attr_macro

Builtin attributes.

Macros

register_builtin 🔒

Enums

BuiltinAttrExpander

Functions

derive_expand 🔒

We generate a very specific expansion here, as we do not actually expand the #[derive] attribute itself in name res, but we do want to expand it to something for the IDE layer, so that the input derive attributes can be downmapped, and resolved as proper paths. This is basically a hack, that simplifies the hacks we need in a lot of ide layer places to somewhat inconsistently resolve derive attributes.

dummy_attr_expand 🔒

dummy_gate_test_expand 🔒

find_builtin_attr

pseudo_derive_attr_expansion