hir_expand

Module db

Defines database & queries for macro expansion.

Structs

ExpandDatabaseData 🔒

SyntaxContextWrapper

Enums

CowArc 🔒

TokenExpander

Constants

TOKEN_LIMIT 🔒

Total limit on the number of tokens produced by any macro invocation.

Traits

ExpandDatabase

Functions

ast_id_map 🔒

check_tt_count 🔒

create_data_ExpandDatabase 🔒

expand_proc_macro 🔒

expand_speculative

This expands the given macro call, but with different arguments. This is used for completion, where we want to see what ‘would happen’ if we insert a token. The token_to_map mapped down into the expansion, with the mapped token(s) returned with their priority.

intern_macro_call 🔒

lookup_intern_macro_call 🔒

macro_arg 🔒

macro_arg_considering_derives 🔒

This resolves the MacroCallId to check if it is a derive macro if so get the macro_arg for the derive. Other wise return the macro_arg for the macro_call_id.

macro_expand 🔒

parse_macro_expansion 🔒

parse_macro_expansion_error 🔒

parse_or_expand 🔒

Main public API – parses a hir file, not caring whether it’s a real file or a macro expansion.

parse_with_map 🔒

proc_macro_span 🔒

resolve_span 🔒

syntax_context 🔒

token_tree_to_syntax_node 🔒

Type Aliases

MacroArgResult 🔒

This is just to ensure the types of smart_macro_arg and macro_arg are the same