hir_expand::db

Function proc_macro_span

fn proc_macro_span(db: &dyn ExpandDatabase, ast: AstId<Fn>) -> Span

Retrieves the span to be used for a proc-macro expansions spans. This is a firewall query as it requires parsing the file, which we don’t want proc-macros to directly depend on as that would cause to frequent invalidations, mainly because of the parse queries being LRU cached. If they weren’t the invalidations would only happen if the user wrote in the file that defines the proc-macro.