Type Alias hir_expand::files::InFile

pub type InFile<T> = InFileWrapper<HirFileId, T>;

Aliased Type

struct InFile<T> {
    pub file_id: HirFileId,
    pub value: T,
}

Fields
file_id: HirFileIdvalue: T
Implementations

impl<SN: Borrow<SyntaxNode>> InFile<SN>

pub fn parent_ancestors_with_macros( self, db: &dyn ExpandDatabase ) -> impl Iterator<Item = InFile<SyntaxNode>> + '_

pub fn ancestors_with_macros( self, db: &dyn ExpandDatabase ) -> impl Iterator<Item = InFile<SyntaxNode>> + '_

pub fn kind(&self) -> SyntaxKind

pub fn text_range(&self) -> TextRange

pub fn original_file_range_rooted(self, db: &dyn ExpandDatabase) -> FileRange

Falls back to the macro call range if the node cannot be mapped up fully.

For attributes and derives, this will point back to the attribute only. For the entire item use [InFile::original_file_range_full].

pub fn original_file_range_with_macro_call_body( self, db: &dyn ExpandDatabase ) -> FileRange

Falls back to the macro call range if the node cannot be mapped up fully.

pub fn original_syntax_node_rooted( self, db: &dyn ExpandDatabase ) -> Option<InRealFile<SyntaxNode>>

impl InFile<&SyntaxNode>

pub fn original_file_range_opt( self, db: &dyn ExpandDatabase ) -> Option<(FileRange, SyntaxContextId)>

Attempts to map the syntax node back up its macro calls.

impl InFile<SyntaxToken>

pub fn original_file_range(self, db: &dyn ExpandDatabase) -> FileRange

Falls back to the macro call range if the node cannot be mapped up fully.

pub fn original_file_range_opt( self, db: &dyn ExpandDatabase ) -> Option<FileRange>

Attempts to map the syntax node back up its macro calls.

impl InFile<TextRange>

pub fn original_node_file_range( self, db: &dyn ExpandDatabase ) -> (FileRange, SyntaxContextId)

pub fn original_node_file_range_rooted( self, db: &dyn ExpandDatabase ) -> FileRange

pub fn original_node_file_range_with_macro_call_body( self, db: &dyn ExpandDatabase ) -> FileRange

pub fn original_node_file_range_opt( self, db: &dyn ExpandDatabase ) -> Option<(FileRange, SyntaxContextId)>

impl<N: AstNode> InFile<N>

pub fn original_ast_node_rooted( self, db: &dyn ExpandDatabase ) -> Option<InRealFile<N>>