hir_expand

Module fixup

To make attribute macros work reliably when typing, we need to take care to fix up syntax errors in the code we’re passing to them.

Structs

SyntaxFixupUndoInfo

This is the information needed to reverse the fixups.

SyntaxFixups 🔒

The result of calculating fixes for a syntax node – a bunch of changes (appending to and replacing nodes), the information that is needed to reverse those changes afterwards, and a token map.

Constants

FIXUP_DUMMY_AST_ID 🔒

FIXUP_DUMMY_RANGE 🔒

FIXUP_DUMMY_RANGE_END 🔒

Functions

can_handle_error 🔒

fixup_syntax 🔒

has_error 🔒

has_error_to_handle 🔒

reverse_fixups 🔒

reverse_fixups_ 🔒