hir_expand

Function map_node_range_up

pub fn map_node_range_up(
    db: &dyn ExpandDatabase,
    exp_map: &ExpansionSpanMap,
    range: TextRange,
) -> Option<(FileRange, SyntaxContext)>

Maps up the text range out of the expansion hierarchy back into the original file its from.

this will return None if any anchors or syntax contexts differ.