hir_expand

Function map_node_range_up_aggregated

pub fn map_node_range_up_aggregated(
    db: &dyn ExpandDatabase,
    exp_map: &ExpansionSpanMap,
    range: TextRange,
) -> FxHashMap<(SpanAnchor, SyntaxContext), TextRange>

Maps up the text range out of the expansion hierarchy back into the original file its from. This version will aggregate the ranges of all spans with the same anchor and syntax context.