hir_expand
Function map_node_range_up_rooted

pub fn map_node_range_up_rooted(
    db: &dyn ExpandDatabase,
    exp_map: &ExpansionSpanMap,
    range: TextRange,
) -> Option<FileRange>

Maps up the text range out of the expansion hierarchy back into the original file its from only considering the root spans contained. Unlike map_node_range_up, this will not return None if any anchors or syntax contexts differ.