hir_expand
Module inert_attr_macro

Builtin attributes resolved by nameres.

The actual definitions were copied from rustc’s compiler/rustc_feature/src/builtin_attrs.rs.

It was last synchronized with upstream commit c3def263a44e07e09ae6d57abfc8650227fb4972.

The macros were adjusted to only expand to the attribute name, since that is all we need to do name resolution, and BUILTIN_ATTRIBUTES is almost entirely unchanged from the original, to ease updating.

Structs

AttributeTemplate

A template that the attribute input must match. Only top-level shape (#[attr] vs #[attr(...)] vs #[attr = ...]) is considered now.

BuiltinAttribute

Constants

INERT_ATTRIBUTES

Attributes that have a special meaning to rustc or rustdoc.

Functions

find_builtin_attr_idx