hir_expand::tt

Type Alias Span

pub type Span = SpanData<SyntaxContext>;

Aliased Type

pub struct Span {
    pub range: TextRange,
    pub anchor: SpanAnchor,
    pub ctx: SyntaxContext,
}

Fields
range: TextRange

The text range of this span, relative to the anchor. We need the anchor for incrementality, as storing absolute ranges will require recomputation on every change in a file at all times.

anchor: SpanAnchor

The anchor this span is relative to.

ctx: SyntaxContext

The syntax context of the span.