Crate hir_expand

hir_expand deals with macro expansion.

Specifically, it implements a concept of MacroFile – a file whose syntax tree originates not from the text of some FileId, but from some macro expansion.

Re-exports

pub use crate::files::AstId;

pub use crate::files::ErasedAstId;

pub use crate::files::FileRange;

pub use crate::files::InFile;

pub use crate::files::InMacroFile;

pub use crate::files::InRealFile;

pub use intern;

Modules

attrs

A higher level attributes based on TokenTree, with also some shortcuts.

builtin

Builtin macros and attributes

change

Defines a unit of change that can applied to the database to get the next state. Changes are transactional.

db

Defines database & queries for macro expansion.

declarative

Compiled declarative macro expanders (macro_rules! and macro)

eager

Eager expansion related utils

files

Things to wrap other things in file ids.

hygiene

Machinery for hygienic macros.

inert_attr_macro

Builtin attributes resolved by nameres.

mod_path

A lowering for use-paths (more generally, paths without angle-bracketed segments).

name

See Name.

proc_macro

Proc Macro Expander stuff

span_map

Span maps for real files and macro expansions.

tt

Macros

__known_path

__path

__tool_path

impl_intern_lookup

quote

FIXME: It probably should implement in proc-macro

Structs

DeclarativeMacro

This struct contains AST for a single macro_rules definition. What might be very confusing is that AST has almost exactly the same shape as tt::TokenTree, but there’s a crucial difference: in macro rules, $ident and $()* have special meaning (see Var and Repeat data structures)

EagerCallInfo

EditionedFileId

ExpandError

ExpansionInfo

ExpansionInfo mainly describes how to map text range between src and expanded macro

MacroCallId

MacroCallLoc

MacroDefId

RenderedExpandError

ValueResult

Enums

ExpandErrorKind

ExpandTo

In Rust, macros expand token trees to token trees. When we want to turn a token tree into an AST node, we need to figure out what kind of AST node we want: something like foo can be a type, an expression, or a pattern.

HirFileId

MacroCallKind

MacroDefKind

MacroKind

Traits

Intern

Lookup

Functions

check_cfg_attr_value

map_node_range_up

Maps up the text range out of the expansion hierarchy back into the original file its from.

map_node_range_up_aggregated

Maps up the text range out of the expansion hierarchy back into the original file its from. This version will aggregate the ranges of all spans with the same anchor and syntax context.

map_node_range_up_rooted

Maps up the text range out of the expansion hierarchy back into the original file its from only considering the root spans contained. Unlike map_node_range_up, this will not return None if any anchors or syntax contexts differ.

prettify_macro_expansion

Inserts whitespace and replaces $crate in macro expansions.

span_for_offset

Looks up the span at the given offset.

Type Aliases

ExpandResult