hir_expand
Module db

Defines database & queries for macro expansion.

Structs

AstIdMapQuery

DeclMacroExpanderQuery

ExpandDatabaseGroupStorage__

ExpandDatabaseStorage

Representative struct for the query group.

ExpandProcMacroQuery

InternMacroCallLookupQuery

InternMacroCallQuery

InternSyntaxContextLookupQuery

InternSyntaxContextQuery

MacroArgQuery

ParseMacroExpansionErrorQuery

ParseMacroExpansionQuery

ProcMacroSpanQuery

ProcMacrosQuery

RealSpanMapQuery

Enums

TokenExpander

Traits

ExpandDatabase

Functions

expand_speculative

This expands the given macro call, but with different arguments. This is used for completion, where we want to see what ‘would happen’ if we insert a token. The token_to_map mapped down into the expansion, with the mapped token(s) returned with their priority.