Module hir_expand::db

Defines database & queries for macro expansion.

Structs

AstDatabaseGroupStorage__

AstDatabaseStorage

Representative struct for the query group.

AstIdMapQuery

ExpandProcMacroQuery

HygieneFrameQuery

InternMacroCallLookupQuery

InternMacroCallQuery

MacroArgTextQuery

MacroDefQuery

MacroExpandErrorQuery

MacroExpandQuery

ParseMacroExpansionQuery

Enums

TokenExpander

Traits

AstDatabase

Functions

expand_speculative

This expands the given macro call, but with different arguments. This is used for completion, where we want to see what ‘would happen’ if we insert a token. The token_to_map mapped down into the expansion, with the mapped token returned.