Module hir_expand::name

See Name.

Re-exports

pub use crate::name;

pub use crate::name;

Modules

known

Macros

name

Structs

Name

Name is a wrapper around string, which is used in hir for both references and declarations. In theory, names should also carry hygiene info, but we are not there yet!

UnescapedName

Wrapper of Name to print the name without “r#” even when it is a raw identifier.

Traits

AsName